Fix overflow issues

+++ src/constants/sprites.rs
@@ -1,4 +1,4 @@
 pub const SPRITE_NUMBER: u8 = 40;
 pub const SPRITE_SIZE_BYTES: u8 = 4;
-pub const SPRITE_X_OFFSET: u8 = 8;
-pub const SPRITE_Y_OFFSET: u8 = 16;
+pub const SPRITE_X_OFFSET: i32 = 8;
+pub const SPRITE_Y_OFFSET: i32 = 16;
